[fe] positional_enconding

+++ supervised_learning/0x11-attention/4-positional_encoding.py
@@ -7,7 +7,6 @@ Positional Encoding
 """
 
 import numpy as np
-import tensorflow as tf
 
 
 def positional_encoding(max_seq_len, dm):
@@ -26,4 +25,4 @@ def positional_encoding(max_seq_len, dm):
             result[i, j] = np.sin(i * div_term)
             result[i, j + 1] = np.cos(i * div_term)
 
-    return tf.cast(result, dtype=tf.float32)
+    return result
